pyproject.toml
[tool.poetry]
name = "langgraph-tutorial"
version = "0.1.0"
description = ""
authors = ["Season Saw <season.saw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
nanoid = "^2.0.0"
langgraph = "^0.0.69"
langsmith = "^0.1.81"
langchain-anthropic = "^0.1.15"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"
langchain = "^0.2.5"
tavily-python = "^0.3.3"
streamlit = "^1.36.0"
langchain-community = "^0.2.5"
watchdog = "^4.0.1"
langchain-nomic = "^0.1.2"
tiktoken = "^0.7.0"
langchainhub = "^0.1.20"
chromadb = "^0.5.3"
langchain-chroma = "^0.1.1"
protobuf = "3.20.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
nomic = {extras = ["local"], version = "^3.0.33"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
